suzerain: fix several GCC 4.4 warnings
Mainly of signed/unsigned comparison type

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@13805 0bd85f0b-c151-43b9-a1e9-27231104551f

+++ suzerain/bspline.c
@@ -678,8 +678,8 @@ compute_banded_collocation_derivative_submatrix(
                                        &dbjstart, &dbjend, bw, dbw);
 
         /* Coerce dbjstart/dbjend to stay within the submatrix of interest */
-        const size_t jstart = GSL_MAX(dbjstart, joffset);
-        const size_t jend   = GSL_MIN(dbjend, npoints-1+joffset);
+        const size_t jstart = GSL_MAX(dbjstart, (size_t) joffset);
+        const int    jend   = GSL_MIN(dbjend, (size_t)(npoints-1+joffset));
 
         for (int k = 0; k <= nderivatives; ++k) {
             for (int j = jstart; j <= jend; ++j) {
